fix: the modal of the square api and payment
