+++ src/component/addressModal/AddressModal.module.scss
@@ -1,0 +1,173 @@
+/*modal address code*/
+.address_modal {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: rgba(0, 0, 0, 0.5);
+  display: flex;
+  z-index: 1000;
+}
+
+.modal_container {
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  width: 400px;
+  height: 80%;
+  background: #fff5ee;
+  display: flex;
+  flex-direction: column;
+  background: #fff5ee;
+  border-radius: 20px;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.2);
+  padding: 1.5rem;
+  margin: 0px;
+}
+
+.addressHeader {
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  margin: 0;
+  padding: 0rem;
+  margin-bottom: 1rem;
+}
+
+.addressHeaderLeft {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  float: left;
+  margin: 0;
+}
+.confirmText {
+  color: #99592a;
+  margin: 0;
+  padding: 0;
+}
+.confirmHeaderText {
+  font-size: 25px;
+  font-weight: 500;
+  margin: 0;
+  padding: 0;
+}
+
+.confirmCloseIcon {
+  font-size: 20px;
+  cursor: pointer;
+  &:disabled {
+    color: #989795;
+  }
+  &:hover,
+  &:active {
+    color: #2a160c;
+  }
+}
+
+.modal_body {
+  margin: 0;
+  display: flex;
+  flex-flow: column;
+  padding: 0;
+  margin-bottom: 1rem;
+}
+
+.ruler {
+  width: 100%;
+  height: 2px;
+  background: #989795;
+  margin-top: 0.5rem;
+}
+
+.title {
+  margin: 0rem;
+  font-weight: bold;
+  font-family: Poppins;
+  font-size: 16px;
+  padding: 0;
+}
+
+.form {
+  .formGroup {
+    display: flex;
+    gap: 1rem;
+    align-items: center;
+    margin-bottom: 1rem;
+    .address_label {
+      font-family: "Poppins";
+      font-style: normal;
+      font-weight: 500;
+      font-size: 14px;
+      margin-bottom: 1rem;
+      /* identical to box height, or 28px */
+      letter-spacing: 0.02em;
+      color: #101010;
+    }
+    .address_input {
+      margin-top: 0.5rem;
+      border-radius: 20px;
+      padding: 7px 13px;
+      border: 1px solid #99592a;
+      width: 100%;
+      text-align: center;
+      color: #101010;
+      height: 40px;
+    }
+  }
+
+  .btn_container {
+    display: flex;
+    margin-top: 2rem;
+
+    .address_back_button {
+      border: #99592a 1px solid;
+      color: #99592a;
+      border-radius: 20px;
+      font-size: 16px;
+      padding: 7px 13px;
+      width: 170px;
+      display: flex;
+      justify-content: center;
+      height: 40px;
+      &:active,
+      &:hover {
+        border: 1px solid #2a160c;
+      }
+      &:disabled {
+        border: 1px solid #989795;
+      }
+    }
+
+    .payment_button {
+      border-radius: 20px;
+ padding: 7px 13px;
+      background: #99592a;
+      border: 1px solid #99592a;
+      width: 170px;
+      display: flex;
+      text-align: center;
+      color: white;
+      justify-content: center;
+      border: none;
+      height: 40px;
+      &:active,
+      &:hover {
+        background: #2a160c;
+      }
+      &:disabled {
+        background: #989795;
+      }
+    }
+  }
+}
+
+.errorMsg {
+  font-size: 13px;
+  color: red;
+  display: flex;
+  align-items: flex-start;
+  text-align: left;
+}
