+++ src/component/square_payment/SquarePayment.module.scss
@@ -10,7 +10,7 @@
 }
 
 .modal_content {
-  width: 35%;
+  width: 350px;
   height: 60%;
   background: #fff5ee;
   display: flex;
